+++ Makefile
@@ -1,6 +1,7 @@
 NAME = minishell
 CC = gcc
 CFLAGS = -Werror -Wall -Wextra -g
+# CFLAGS += -fsanitize=address
 LINK_LIB = -lcurses
 ifeq ($(shell uname), Linux)
 	LINK_LIB += -lbsd
@@ -11,8 +12,6 @@ LIBFT_MAKE = $(MAKE) -C $(LIBFT_PATH)
 LIBFT_LIB = -L./libft -lft
 
 LDFLAGS = $(LIBFT_LIB) $(LINK_LIB)
-CFLAGS = -g
-# CFLAGS = -g -fsanitize=address
 
 HEADER_FILES = minishell.h
 
@@ -25,8 +24,9 @@ SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
 	builtin_echo.c builtin_env.c builtin_exit.c builtin_cd.c				\
 	builtin_cd_path.c builtin_cd_chdir.c builtin_cd_cdpath.c				\
 	builtin_export.c builtin_pwd.c builtin_unset.c str_utils.c				\
-	shell_initialization.c input.c rope1.c rope2.c rope3.c rope4.c			\
-	splay1.c splay2.c splay3.c splay4.c
+	shell_initialization.c rope1.c rope2.c rope3.c rope4.c splay1.c			\
+	splay2.c splay3.c splay4.c editor1.c editor2.c editor3.c				\
+	editor4.c editor5.c
 
 OBJS = $(SRCS:.c=.o)
