+++ minishell.c
@@ -2,7 +2,7 @@
 #include "libft/libft.h"
 #include "minishell.h"
 #include "parse.h"
-#include "input.h"
+#include "editor.h"
 
 void	die(void)
 {
@@ -74,30 +74,8 @@ int	do_command(char *cmdstr)
 
 int	main(int argc, char **argv)
 {
-	/* t_parse_ast				*cmdline; */
-	/* t_parse_buffer			buf; */
-	/* t_token					tok; */
-	/* t_parse_ast				*seqcmd; */
-
 	initialize_shell();
 	if (argc == 3 && ft_strncmp(argv[1], "-c", 3) == 0)
 		return (do_command(argv[2]));
-	/* init_buffer_with_string(&buf, ""); */
-	/* printf("Welcome to Minishell\n"); */
-	/* while (1) */
-	/* { */
-	/* 	printf("\r"PROMPT); */
-	/* 	fflush(stdout); */
-	/* 	lex_get_token(&buf, &tok); */
-	/* 	cmdline = parse_command_line(&buf, &tok); */
-	/* 	if (cmdline) */
-	/* 	{ */
-	/* 		seqcmd = cmdline->content.command_line->seqcmd_node; */
-	/* 		invoke_sequential_commands(seqcmd); */
-	/* 		parse_free_all_ast(); */
-	/* 	} */
-	/* 	else */
-	/* 		put_err_msg("Parse error."); */
-	/* } */
 	edit_main();
 }
